# internal/api/server.gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.0 DO NOT EDIT.
package api

import (
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

// Defines values for Currency.
const (
	EurBcv      Currency = "eur_bcv"
	UsdBcv      Currency = "usd_bcv"
	UsdtBinance Currency = "usdt_binance"
)

// Valid indicates whether the value is a known member of the Currency enum.
func (e Currency) Valid() bool {
	switch e {
	case EurBcv:
		return true
	case UsdBcv:
		return true
	case UsdtBinance:
		return true
	default:
		return false
	}
}

// Defines values for HealthStatus.
const (
	Degraded HealthStatus = "degraded"
	Ok       HealthStatus = "ok"
)

// Valid indicates whether the value is a known member of the HealthStatus enum.
func (e HealthStatus) Valid() bool {
	switch e {
	case Degraded:
		return true
	case Ok:
		return true
	default:
		return false
	}
}

// AllRates defines model for AllRates.
type AllRates struct {
	EurBcv      RateEntry `json:"eur_bcv"`
	UsdBcv      RateEntry `json:"usd_bcv"`
	UsdtBinance RateEntry `json:"usdt_binance"`
}

// Currency defines model for Currency.
type Currency string

// Error defines model for Error.
type Error struct {
	Error string `json:"error"`
}

// Health defines model for Health.
type Health struct {
	// Stale List of currencies whose data exceeded the staleness threshold.
	Stale  *[]Currency  `json:"stale,omitempty"`
	Status HealthStatus `json:"status"`
}

// HealthStatus defines model for Health.Status.
type HealthStatus string

// HistoryEntry defines model for HistoryEntry.
type HistoryEntry struct {
	RecordedAt time.Time `json:"recorded_at"`
	Value      float64   `json:"value"`
}

// RateEntry defines model for RateEntry.
type RateEntry struct {
	DataAgeSeconds int        `json:"data_age_seconds"`
	IsStale        bool       `json:"is_stale"`
	LastUpdated    *time.Time `json:"last_updated"`
	Value          float64    `json:"value"`
}

// RateHistory defines model for RateHistory.
type RateHistory struct {
	Currency Currency       `json:"currency"`
	Entries  []HistoryEntry `json:"entries"`
}

// GetRatesCurrencyHistoryParams defines parameters for GetRatesCurrencyHistory.
type GetRatesCurrencyHistoryParams struct {
	// FromDate Start date (inclusive) in YYYY-MM-DD format
	FromDate string `form:"fromDate" json:"fromDate"`

	// ToDate End date (inclusive) in YYYY-MM-DD format
	ToDate string `form:"toDate" json:"toDate"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /health)
	GetHealth(w http.ResponseWriter, r *http.Request)

	// (GET /rates)
	GetRates(w http.ResponseWriter, r *http.Request)

	// (GET /rates/{currency})
	GetRatesCurrency(w http.ResponseWriter, r *http.Request, currency Currency)

	// (GET /rates/{currency}/history)
	GetRatesCurrencyHistory(w http.ResponseWriter, r *http.Request, currency Currency, params GetRatesCurrencyHistoryParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetRates operation middleware
func (siw *ServerInterfaceWrapper) GetRates(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetRates(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetRatesCurrency operation middleware
func (siw *ServerInterfaceWrapper) GetRatesCurrency(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "currency" -------------
	var currency Currency

	err = runtime.BindStyledParameterWithOptions("simple", "currency", r.PathValue("currency"), &currency, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "currency", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetRatesCurrency(w, r, currency)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetRatesCurrencyHistory operation middleware
func (siw *ServerInterfaceWrapper) GetRatesCurrencyHistory(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "currency" -------------
	var currency Currency

	err = runtime.BindStyledParameterWithOptions("simple", "currency", r.PathValue("currency"), &currency, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "currency", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetRatesCurrencyHistoryParams

	// ------------- Required query parameter "fromDate" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "fromDate", r.URL.Query(), &params.FromDate, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "fromDate"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "fromDate", Err: err})
		}
		return
	}

	// ------------- Required query parameter "toDate" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "toDate", r.URL.Query(), &params.ToDate, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "toDate"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "toDate", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetRatesCurrencyHistory(w, r, currency, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/health", wrapper.GetHealth)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/rates", wrapper.GetRates)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/rates/{currency}", wrapper.GetRatesCurrency)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/rates/{currency}/history", wrapper.GetRatesCurrencyHistory)

	return m
}